now using geom.json
cleaned up lotsof compiler warnings
mroe sensible use of json_module.mod and libjsonfortran.a

+++ src/XSPH/szlz.f
@@ -103,7 +103,7 @@ c     Start the cycle over energy points (ie)
          endif
 
          call fixdsx (iph, dx, rgrd , dgc, dpc, dgcn, dpcn)
-        jri = (log(rmt(iph)) + x0) / rgrd + 2
+        jri = int((log(rmt(iph)) + x0) / rgrd) + 2
         jri1 = jri+1
         eref = vtotph(jri1)
         do 40 i = 1, jri1
